Initial commit
Game is parsed and stored in the `game` struct. The actual "gameplay"
is stored in `state` structure which only references the game via read-only
access somewhat as living in the same package we can still access private members.

BFS approach was taken with a fuse of max score as we can visit same nodes
and it's hard to tell that we aren't going in loops needlessly. If we would go
always in 1x1 increments then visiting same node would be suboptimal and
we could mark it as visited, in hopper due to it's velocity jumping nature
I currently see no way to optimise hence the fuse added.

Games are decoded from games.txt in a format(line by line):
- number of games
- grid size
- startpos, finishpos
- obstacle groups
- for each obstacle line represents x1,x2 and y1,y2 contained block of obstacles